# Make sideof(::Point, ::Ring) serial when convenient