--- a/src/sideof.jl
+++ b/src/sideof.jl
@@ -72,7 +72,7 @@ function _sideof_hao_threads(point::Point, r::Ring)
   k = Threads.Atomic{Int}(0)
   ison = Threads.Atomic{Bool}(false)
   Threads.@threads for i in eachindex(v)
-    _ison, increment_k = _sideof_hao_core(point, v[i], v[i+1])
+    _ison, increment_k = _sideof_hao_core(point, v[i], v[i + 1])
     (ison[] = _ison) && break
     increment_k && Threads.atomic_add!(k, 1)
   end
@@ -83,7 +83,7 @@ function _sideof_hao_serial(point::Point, r::Ring)
   v = vertices(r)
   k = 0
   for i in eachindex(v)
-    ison, increment_k = _sideof_hao_core(point, v[i], v[i+1])
+    ison, increment_k = _sideof_hao_core(point, v[i], v[i + 1])
     ison && return ON
     k += increment_k
   end
@@ -105,7 +105,7 @@ function _sideof_hao_core(point::Point, pᵢ::Point, pⱼ::Point)
   pⱼ = flat(coords(pⱼ))
   xᵢ, yᵢ = pᵢ.x, pᵢ.y
   xⱼ, yⱼ = pⱼ.x, pⱼ.y
-  
+
   v₁ = yᵢ - yₚ
   v₂ = yⱼ - yₚ
 
@@ -121,47 +121,47 @@ function _sideof_hao_core(point::Point, pᵢ::Point, pⱼ::Point)
     # case 3, 9, 16, 21, 13, 24
     f = u₁ * v₂ - u₂ * v₁
     if ispositive(f)
-    # case 3, 9
+      # case 3, 9
       return INCREMENT_K
     elseif isequalzero(f)
-    # case 16, 21
+      # case 16, 21
       return ISON
     end
   elseif ispositive(v₁) && isnonpositive(v₂)
     # case 4, 10, 19, 20, 12, 25
     f = u₁ * v₂ - u₂ * v₁
     if isnegative(f)
-    # case 4, 10
+      # case 4, 10
       return INCREMENT_K
     elseif isequalzero(f)
-    # case 19, 20
+      # case 19, 20
       return ISON
     end
   elseif isequalzero(v₂) && isnegative(v₁)
     # case 7, 14, 17
     f = u₁ * v₂ - u₂ * v₁
     if isequalzero(f)
-    # case 17
+      # case 17
       return ISON
     end
   elseif isequalzero(v₁) && isnegative(v₂)
     # case 8, 15, 18
     f = u₁ * v₂ - u₂ * v₁
     if isequalzero(f)
-    # case 18
+      # case 18
       return ISON
     end
   elseif isequalzero(v₁) && isequalzero(v₂)
     # case 1, 2, 5, 6, 22, 23
     if isnonpositive(u₂) && isnonnegative(u₁)
-    # case 1
+      # case 1
       return ISON
     elseif isnonpositive(u₁) && isnonnegative(u₂)
-    # case 2
+      # case 2
       return ISON
     end 
 end
-# case 5, 6, 7, 8, 12, 13, 14, 15, 22, 23, 24, 25
+  # case 5, 6, 7, 8, 12, 13, 14, 15, 22, 23, 24, 25
   return UNCHANGED
 end