--- a/src/sideof.jl
+++ b/src/sideof.jl
@@ -159,12 +159,11 @@ function _sideof_hao_core(point::Point, pᵢ::Point, pⱼ::Point)
     elseif isnonpositive(u₁) && isnonnegative(u₂)
       # case 2
       return ISON
-    end 
+    end
   end
   # case 5, 6, 7, 8, 12, 13, 14, 15, 22, 23, 24, 25
   return UNCHANGED
 end
-
 
 # -----
 # MESH