--- a/test/sideof.jl
+++ b/test/sideof.jl
@@ -65,7 +65,7 @@
 
   # sideof serial vs parallel
   r = randpoint2(500) |> Ring
-  p = Point(2,2)
+  p = Point(2, 2)
   serial = Meshes._sideof_hao_serial(p, r)
   parallel = Meshes._sideof_hao_threads(p, r)
   @test serial === parallel