@disberd 2024-08-08
I am not sure which tests to add, I think correctness was already covered by previous tests, maybe we could just add a test with a ring of <= 1000 points and test that this does not allocate. 
I am open to suggestions for tests