@juliohm 2024-08-09
It looks good to me. Let's wait for another pair of eyes given that this change is very sensitive. Waiting for your input @eliascarv 👍🏽